requirements.txt


langchain==0.3.25
langchain-google-genai==2.1.4
langchain-core==0.3.58
langgraph==0.4.1
python-dotenv==1.1.0
